requirements.txt
aiofiles==24.1.0
aiogram==3.21.0
aiogram_i18n==1.4
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
alembic==1.16.4
annotated-types==0.7.0
async-timeout==5.0.1
asyncpg==0.30.0
attrs==25.3.0
certifi==2025.7.14
click==8.1.8
colorama==0.4.6
exceptiongroup==1.3.0
frozenlist==1.7.0
greenlet==3.2.3
idna==3.10
iniconfig==2.1.0
magic-filter==1.0.12
Mako==1.3.10
MarkupSafe==3.0.2
multidict==6.6.3
packaging==25.0
pluggy==1.6.0
propcache==0.3.2
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pytest==8.4.1
pytest-asyncio==1.0.0
python-dotenv==1.1.1
redis==6.2.0
SQLAlchemy==2.0.41
tomli==2.2.1
typing-inspection==0.4.1
typing_extensions==4.14.1
yarl==1.20.1
fastapi==0.104.1
uvicorn==0.24.0